Extending and refining EOQ+ Newsvendor and Profit.

+++ backend/inventorysimulator/src/main/java/com/retailops/inventorysimulator/simulator/dto/SimulationRunDTO.java
@@ -3,14 +3,16 @@ package com.retailops.inventorysimulator.simulator.dto;
 
 import com.retailops.inventorysimulator.util.SimulationType;
 
+import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.time.LocalDateTime;
 
 public record SimulationRunDTO(Long id,
                                SimulationType simulationType,
                                String productName,
-                               int stockQty,
-                               int demand,
-                               double profit,
+                               BigInteger stockQty,
+                               BigInteger demand,
+                               BigDecimal profit,
                                LocalDateTime runAt,
                                String username) {
 }
